Refactor Price Providers
